I accidentally used MapDemo to be about map the collection and map the function, so I split out map the collection into its own thing.
I also added a pattern match worksheet.
